fix: Rename to ChangePasswordModal
